requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pytest==8.3.4
beautifulsoup4==4.13.4
lxml==5.4.0
ftfy==6.3.1
pymongo==4.16.0
python-dotenv==1.2.2
google-genai
